pyproject.toml
[tool.poetry]
name = "embedding-lenses"
version = "0.1.0"
description = ""
authors = ["edugp <edugp91@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
huggingface-hub = "0.0.19"
streamlit = "1.1.0"
transformers = "4.11.3"
watchdog = "2.1.3"
sentence-transformers = "2.0.0"
bokeh = "2.2.2"
numpy = "1.20.0"
numba = "^0.54.1"
umap-learn = "^0.5.2"
datasets = "1.14.0"
black = "^21.10b0"
flake8 = "^4.0.1"
scikit-learn = "0.24.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.black]
line-length = 160
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 160
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
